requirements.txt
flask
sentence-transformers
faiss-cpu
transformers
torch
requests
python-dotenv
